Automated versioning and publishing to NPM
There are a lot of steps involved in publishing a new version of a
package to NPM:
- Tagging a new version;
- Pushing the code to GitHub
- Publishing the package to NPM
etc...

I came across this article from Isaac Z. Schlueter (NPM)
[Easy Automatic npm Publishes]
(https://blog.izs.me/2019/04/easy-automatic-npm-publishes)
a couple of months ago and although I'm of the opinion that
publishing's better done from a CI/CD tool (such as Azure Pipelines
or Travis), this is better than what I have today, which is nothing
at all.
I'm not really in the mood to figure out how to do it from any of
those CI/CD right now, so in the spirit of [Always be Automating]
(https://queue.acm.org/detail.cfm?id=3197520), this revision
automates versioning and publishing from a local dev's box.